requirements.txt
# Core packages
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.10
langgraph>=0.0.20
langsmith>=0.0.69

# OpenAI
openai>=1.3.0

# Vector database
chromadb>=0.4.18
sentence-transformers>=2.2.2
pypdf>=3.17.1

# Document loading and processing
beautifulsoup4>=4.12.2
lxml>=4.9.3
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
pydantic>=2.5.2
typing-extensions>=4.8.0
tqdm>=4.66.1 
certifi>=2024.2.2

# Web interface
streamlit>=1.31.0

# ##
# pip install langserve fastapi uvicorn
# pip install sse_starlette
#pip install -qU langchain-google-genai